requirements.txt
# 데이터 수집 및 처리
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0

# 머신러닝
scikit-learn>=1.3.0
xgboost>=1.7.0

# 기술적 분석
ta>=0.10.2

# 시각화
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# 웹 대시보드
flask>=2.3.0
flask-login>=0.6.0
streamlit>=1.25.0

# 자동매매 API
alpaca-trade-api>=3.0.0

# 알림
python-telegram-bot>=20.0
# smtplib는 Python 내장 라이브러리

# 서버 배포
gunicorn>=21.0.0
# nginx는 시스템 패키지 (apt install nginx)

# 유틸리티
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
psutil>=5.9.0
pytz>=2024.1  # 시간대 및 썸머타임 처리
pandas-market-calendars>=4.4.0

# 데이터베이스 (선택사항)
# sqlite3는 Python 내장 라이브러리

# 보고서 생성
jinja2>=3.1.0
weasyprint>=59.0

# AI 레포트 생성
google-generativeai>=0.3.0

# 뉴스 감성 분석
finnhub-python>=2.4.19
textblob>=0.17.1
transformers>=4.30.0
torch>=2.0.0

# 거시경제 지표
fredapi>=0.5.1

